# Project/docs/a9_openapi.yaml
openapi: 3.0.0
info:
 version: '2.0'
 title: 'OnlyFEUP Web API'
 description: 'Web Resources Specification (A9) for OnlyFEUP'
servers:
- url: https://lbaw2255.lbaw.fe.up.pt
  description: Production server
...
paths:
...
  /api/group:

    get:
      operationId: R405
      summary: 'R405 : Search groups'
      description: 'Search groups. Access: USR, ADM'
      tags:
        - 'M04: Search'

      parameters:
        - in: query
          name: content
          description: 'Search content'
          schema:
            type: string
          required: true

      responses:
        '200':
          description: 'Success. Returns some HTML text containing a list of groups information'
        '403':
          description: 'Forbiden action. You need to be logged in first'

############################################ ADMINISTRATION ############################################

######### ADMIN PAGE #########
...
  /group/{id}/edit:

    get:
      operationId: R703
      summary: 'R703: Edit group'
      description: 'Shows the edit group page. Access: OWN'
      tags:
        - 'M07: Groups'

      parameters:
      - in: query
        name: id
        description: 'Group ID'
        schema:
          type: integer
        required: true

      responses:
        '200':
          description: 'Ok. You can now edit. Show edit group UI.'
        '401':
          description: 'Unauthorized. You cannot edit this group.'
          headers:
            Location:
              schema:
                type: string
              examples:
                401Success:
                  description: 'Unauthorized. Redirect to group page.'
                  value: '/group/{id}'

################## GROUP EDIT ####################
  /group/edit:
    
    post:
      operationId: R704
      summary: 'R704: Edit group action'
      description: 'Processes and saves the changes made by group owner. Access: OWN'
      tags:
        - 'M07: Groups'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                  format: password
                image:
                  type: string
                  format: binary
                is_public:
                  type: boolean

      responses:
        '302':
          description: 'Redirect after processing the new group information.'
          headers:
            Location:
              schema:
                type: string
              example:
                302Success:
                  description: 'Successful update. Redirect to group page.'
                  value: '/group/{id}'
                302Failure:
                  description: 'Failed update. Redirect again to edit group page.'
                  value: '/group/{id}/edit'

################## GROUP CREATE ####################
...